requirements.txt
streamlit==1.29.0
streamlit-option-menu==0.3.6
plotly==5.18.0
pandas
openpyxl==3.1.2
python-dotenv==1.0.0
langchain>=0.0.300
langchain-openai>=0.0.2
openai>=0.27.8
tiktoken>=0.4.0
Jinja2==3.1.2
transformers>=4.36.2
huggingface-hub>=0.16.4
langchain-huggingface
pydantic>=2.0.0
Pillow>=10.0.0
torch>=2.0.0
accelerate>=0.20.0
